Remove iconID, fix incorrect icon, refactor function into functional component, approve VR based on consistent SVG width, & remove SVG gulp task.
