+++ package.json
@@ -8,7 +8,7 @@
     "build:dev": "npm run remove-dist && npm run build:static && webpack --mode=development --debug --devtool cheap-source-map --output-pathinfo",
     "build:production": "npm run remove-dist && npm run build:static && webpack --mode=production",
     "build:test": "npm run build:production -- --include-tests",
-    "build:static": "gulp svg imagemin",
+    "build:static": "gulp imagemin",
     "build:storybook": "build-storybook -c .storybook -o dist && npm run build:static",
     "remove-dist": "rimraf ./dist",
     "prezip": "composer run autoload-includes",
@@ -178,7 +178,6 @@
     "git-repo-info": "^2.1.1",
     "gulp": "^4.0.2",
     "gulp-imagemin": "^7.1.0",
-    "gulp-svgmin": "^3.0.0",
     "gulp-zip": "^5.0.2",
     "husky": "^3.1.0",
     "jest-localstorage-mock": "^2.4.2",
